Mipmap shader oncelock Fixes #251
