+++ crates/bitang/src/render/image.rs
@@ -4,7 +4,7 @@ use serde::Deserialize;
 use std::sync::{Arc, RwLock};
 use wgpu::Extent3d;
 
-use super::Size2D;
+use super::{mipmap_generator::MipmapGenerator, Size2D};
 
 #[derive(Debug, Deserialize, Clone, Copy)]
 pub enum PixelFormat {
@@ -160,7 +160,7 @@ impl BitangImage {
         if mip_level_count > 1 {
             let mut command_encoder =
                 context.device.create_command_encoder(&wgpu::CommandEncoderDescriptor::default());
-            generate_mipmaps(&mut command_encoder, &context.device, &texture);
+            MipmapGenerator::generate(&mut command_encoder, &context.device, &texture);
             context.queue.submit(Some(command_encoder.finish()));
         }
 
@@ -326,32 +326,20 @@ impl BitangImage {
     }
 
     pub fn generate_mipmaps(&self, context: &mut FrameContext) -> Result<()> {
-        match &self.inner {
-            ImageInner::Attachment(attachment) => {
-                let attachment = attachment.read().unwrap();
-                let Some(texture) = &attachment.texture else {
-                    bail!("Image '{}' has no texture", self.id);
-                };
-                generate_mipmaps(
-                    &mut context.command_encoder,
-                    &context.gpu_context.device,
-                    texture,
-                );
-                Ok(())
-            }
-            ImageInner::Immutable(texture) => {
-                // TODO: does this case ever run?
-                generate_mipmaps(
-                    &mut context.command_encoder,
-                    &context.gpu_context.device,
-                    texture,
-                );
-                Ok(())
-            }
-            ImageInner::Swapchain(_) => {
-                bail!("Can't generate mipmaps for swapchain image");
-            }
-        }
+        let ImageInner::Attachment(attachment) = &self.inner else {
+            bail!("Mipmaps can only be generated for attachment textures during rendering, and image '{}' is not an attachment texture.", self.id);
+        };
+
+        let attachment = attachment.read().unwrap();
+        let Some(texture) = &attachment.texture else {
+            bail!("Image '{}' has no texture", self.id);
+        };
+        MipmapGenerator::generate(
+            &mut context.command_encoder,
+            &context.gpu_context.device,
+            texture,
+        );
+        Ok(())
     }
 
     pub fn copy_attachment_to_buffer(
@@ -389,107 +377,5 @@ impl BitangImage {
             },
         );
         Ok(())
-    }
-}
-
-fn generate_mipmaps(
-    encoder: &mut wgpu::CommandEncoder,
-    device: &wgpu::Device,
-    texture: &wgpu::Texture,
-) {
-    let mip_count = texture.mip_level_count();
-
-    // TODO: use OnceLock instead of loading this shader each time
-    let shader = device.create_shader_module(wgpu::include_wgsl!("blit.wgsl"));
-    let texture_format = texture.format();
-
-    let pipeline = device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
-        label: Some("blit"),
-        layout: None,
-        vertex: wgpu::VertexState {
-            module: &shader,
-            entry_point: Some("vs_main"),
-            compilation_options: Default::default(),
-            buffers: &[],
-        },
-        fragment: Some(wgpu::FragmentState {
-            module: &shader,
-            entry_point: Some("fs_main"),
-            compilation_options: Default::default(),
-            targets: &[Some(texture_format.into())],
-        }),
-        primitive: wgpu::PrimitiveState {
-            topology: wgpu::PrimitiveTopology::TriangleList,
-            ..Default::default()
-        },
-        depth_stencil: None,
-        multisample: wgpu::MultisampleState::default(),
-        multiview: None,
-        cache: None,
-    });
-
-    let bind_group_layout = pipeline.get_bind_group_layout(0);
-
-    let sampler = device.create_sampler(&wgpu::SamplerDescriptor {
-        label: Some("mip"),
-        address_mode_u: wgpu::AddressMode::ClampToEdge,
-        address_mode_v: wgpu::AddressMode::ClampToEdge,
-        address_mode_w: wgpu::AddressMode::ClampToEdge,
-        mag_filter: wgpu::FilterMode::Linear,
-        min_filter: wgpu::FilterMode::Linear,
-        mipmap_filter: wgpu::FilterMode::Nearest,
-        ..Default::default()
-    });
-
-    let views = (0..mip_count)
-        .map(|mip| {
-            texture.create_view(&wgpu::TextureViewDescriptor {
-                label: Some("mip"),
-                format: None,
-                dimension: None,
-                usage: None,
-                aspect: wgpu::TextureAspect::All,
-                base_mip_level: mip,
-                mip_level_count: Some(1),
-                base_array_layer: 0,
-                array_layer_count: None,
-            })
-        })
-        .collect::<Vec<_>>();
-
-    for target_mip in 1..mip_count as usize {
-        let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
-            layout: &bind_group_layout,
-            entries: &[
-                wgpu::BindGroupEntry {
-                    binding: 0,
-                    resource: wgpu::BindingResource::TextureView(&views[target_mip - 1]),
-                },
-                wgpu::BindGroupEntry {
-                    binding: 1,
-                    resource: wgpu::BindingResource::Sampler(&sampler),
-                },
-            ],
-            label: None,
-        });
-
-        let mut rpass = encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
-            label: None,
-            color_attachments: &[Some(wgpu::RenderPassColorAttachment {
-                view: &views[target_mip],
-                resolve_target: None,
-                ops: wgpu::Operations {
-                    load: wgpu::LoadOp::Clear(wgpu::Color::WHITE),
-                    store: wgpu::StoreOp::Store,
-                },
-            })],
-            depth_stencil_attachment: None,
-            timestamp_writes: None,
-            occlusion_query_set: None,
-        });
-
-        rpass.set_pipeline(&pipeline);
-        rpass.set_bind_group(0, &bind_group, &[]);
-        rpass.draw(0..3, 0..1);
     }
 }
